fix(ui): remove mimetype checks for file uploads
Determining mimetypes of uploaded files is unfortunately not reliable cross-platform.
A file of type 'application/x-shellscript` in Firefox on macos, may have a filetype of
"" on Firefox in linux. Given this has been a source of bugs, enforcing a mimetype for
script uploads is probably more of a hindrance than a help, hence removing this feature
entirely here.

Fixes #1666 lp#1896179